requirements.txt
tk
google-cloud-vision
google-cloud-translate
opencv-python
Pillow
deepl 